Incremental checkin.
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@1411 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ sandbox/rick/nolong/kernel/platform/windows/ExternalFunctions.cpp
@@ -99,7 +99,7 @@
 extern HANDLE apiProtect;
 
 typedef struct _ENVENTRY {                  /* setlocal/endlocal structure    */
-  ULONG    DriveNumber;                     /* saved drive                    */
+  size_t   DriveNumber;                     /* saved drive                    */
   char     Directory[DIRLEN];               /* saved current directory        */
   char    *Environment;                     /* saved environment segment      */
   char     Variables[1];                    /* start of variable values       */
