+++ sandbox/rick/nolong/kernel/platform/windows/StoredMetaData.cpp
@@ -110,7 +110,7 @@ typedef struct _control {              /* meta date control info            */
     unsigned short MetaVersion;        /* version of the meta prog          */
     char           RexxVersion[40];    /* version of rexx intrpreter        */
     FILESTATUS     FileStatus;         /* file information                  */
-    LONG           ImageSize;          /* size of the method info           */
+    size_t         ImageSize;          /* size of the method info           */
 } FILE_CONTROL;                        /* saved control info                */
 
 
@@ -489,8 +489,8 @@ RexxMethod *SysRestoreTranslatedProgram(
   FILE_CONTROL  Control;               /* control information               */
   char         *StartPointer;          /* start of buffered method          */
   RexxBuffer   *Buffer;                /* Buffer to unflatten               */
-  LONG          BufferSize;            /* size of the buffer                */
-  ULONG         BytesRead;             /* actual bytes read                 */
+  size_t        BufferSize;            /* size of the buffer                */
+  size_t        BytesRead;             /* actual bytes read                 */
   RexxMethod   *Method;                /* unflattened method                */
   RexxSource   *Source;                /* REXX source object                */
   RexxActivity *activity;              /* the current activity              */
